Redoing parser again
* Tiles.Content.Bridge.DfNet.IntegrationTests
	This is where all integration tests needed to show that the df
	content bridge works.  I have moved the existing integration tests
	here and created a static content store for the new project.  Now
	it only needs to be read from disk one time, and saves us lots of
	time.

	I have moved all the df raws under source control into this
	directory as well as the existing parser integration test (soon to
	be phased out)

* Tiles.Content
	This is just a bunch of poco models that will ultimately drive the
	engine.  It is modeled separately from persistence, engine
	mechanics, and everything else.

* Tiles.Content.Bridge.DfNet
	This is the content bridge.  It takes the DfNet.Raws api and
	fabricates model objects from Tiles.Content

Later there will be a serialization project for the content, but for now,
it just exists as in-memory model.  That is, until we have the fabrication
and engine mechanics worked out sufficiently well.

The old DwarfFortressNet projects will soon go away as I integrate and
regain functionality in Tiles.Content.Bridge.DfNet

The bridge integration work is being driven by DfAgentFactoryTests.