DF creatures injected as agents
This is the first version where both df weapons and df creatures are
running inside the simulation. It is a little rough around the edges but
it is a good starting point.

The engine treats body attacks as no different from weapon attacks.  The
protoype agents had special weapons embedded in their body parts, so the
combat engine treated them the same way.

We want to be able discover a set of combat moves attached to a body part
that represent unarmed attacks that are possible if that body part is
functional.  Save game before that.